viikko4, tehtävät 4-10
