--- viikko4/LoginWeb2/src/main/java/ohtu/Tester.java
@@ -1,31 +1,0 @@
-package ohtu;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.htmlunit.HtmlUnitDriver;
-
-public class Tester {
-    public static void main(String[] args) {
-        WebDriver driver = new HtmlUnitDriver();
-
-        driver.get("http://localhost:8090");
-        System.out.println( driver.getPageSource() );
-        WebElement element = driver.findElement(By.linkText("login"));       
-        element.click(); 
-        
-        System.out.println("==");
-        
-        System.out.println( driver.getPageSource() );
-        element = driver.findElement(By.name("username"));
-        element.sendKeys("pekka");
-        element = driver.findElement(By.name("password"));
-        element.sendKeys("akkep");
-        element = driver.findElement(By.name("login"));
-        element.submit();
-        
-        System.out.println("==");
-        System.out.println( driver.getPageSource() );
-        
-    }
-}
