+++ viikko4/LoginWeb2/src/test/easyb/logging_in.story
@@ -7,19 +7,19 @@ description 'User can log in with valid username/password-combination'
 
 scenario "user can login with correct password", {
     given 'login selected', {
-        driver = new HtmlUnitDriver();
-        driver.get("http://localhost:8090");
-        element = driver.findElement(By.linkText("login"));       
-        element.click();       
+        driver = new HtmlUnitDriver()
+        driver.get("http://localhost:8090")
+        element = driver.findElement(By.linkText("login"))       
+        element.click()       
     }
 
     when 'a valid username and password are given', {
-        element = driver.findElement(By.name("username"));
-        element.sendKeys("pekka");
-        element = driver.findElement(By.name("password"));
-        element.sendKeys("akkep");
-        element = driver.findElement(By.name("login"));
-        element.submit();
+        element = driver.findElement(By.name("username"))
+        element.sendKeys("pekka")
+        element = driver.findElement(By.name("password"))
+        element.sendKeys("akkep")
+        element = driver.findElement(By.name("login"))
+        element.submit()
     }
  
     then 'user will be logged in to system', {
@@ -28,13 +28,45 @@ scenario "user can login with correct password", {
 }
 
 scenario "user can not login with incorrect password", {
-    given 'command login selected'
-    when 'a valid username and incorrect password are given'
-    then 'user will not be logged in to system'
+    given 'command login selected', {
+        driver = new HtmlUnitDriver()
+        driver.get("http://localhost:8090")
+        element = driver.findElement(By.linkText("login"))     
+        element.click()
+    }
+
+    when 'a valid username and incorrect password are given', {
+        element = driver.findElement(By.name("username"))
+        element.sendKeys("pekka")
+        element = driver.findElement(By.name("password"))
+        element.sendKeys("akke")
+        element = driver.findElement(By.name("login"))
+        element.submit()
+    }
+
+    then 'user will not be logged in to system', {
+        driver.getPageSource().contains("wrong username or password").shouldBe true
+    }
 }
 
 scenario "nonexistent user can not login to system", {
-    given 'command login selected'
-    when 'a nonexistent username and some password are given'
-    then 'user will not be logged in to system'
+    given 'command login selected', {
+        driver = new HtmlUnitDriver()
+        driver.get("http://localhost:8090")
+        element = driver.findElement(By.linkText("login"))       
+        element.click()
+    }
+
+    when 'a nonexistent username and some password are given', {
+        element = driver.findElement(By.name("username"))
+        element.sendKeys("pekk")
+        element = driver.findElement(By.name("password"))
+        element.sendKeys("akkep")
+        element = driver.findElement(By.name("login"))
+        element.submit()
+    }
+
+    then 'user will not be logged in to system', {
+        driver.getPageSource().contains("wrong username or password").shouldBe true
+    }
 }
